refactor: 메일 관련 클래스 패키지 구조 변경 #75
