+++ src/main/java/KUSITMS/WITHUS/global/infra/email/sender/MailSender.java
@@ -1,4 +1,4 @@
-package KUSITMS.WITHUS.global.infra.email;
+package KUSITMS.WITHUS.global.infra.email.sender;
 
 public interface MailSender {
     void send(String to, String subject, String text);
