+++ src/main/java/KUSITMS/WITHUS/global/infra/email/sender/SmtpMailSender.java
@@ -1,4 +1,4 @@
-package KUSITMS.WITHUS.global.infra.email;
+package KUSITMS.WITHUS.global.infra.email.sender;
 
 import KUSITMS.WITHUS.global.exception.CustomException;
 import KUSITMS.WITHUS.global.exception.ErrorCode;
