+++ src/main/java/KUSITMS/WITHUS/global/infra/email/template/MailTemplateProvider.java
@@ -1,4 +1,4 @@
-package KUSITMS.WITHUS.global.infra.email;
+package KUSITMS.WITHUS.global.infra.email.template;
 
 import KUSITMS.WITHUS.global.exception.CustomException;
 import KUSITMS.WITHUS.global.exception.ErrorCode;
